srcs/backend/requirements.txt
Django==5.1.4
python-dotenv==1.0.1
psycopg==3.2.3
djangorestframework==3.15.2
django-filter==24.3
Pygments==2.18.0
django-cors-headers==4.6.0
django-redis==5.4.0
requests==2.25.1
drf-spectacular==0.28.0
dj-rest-auth==7.0.1
django-allauth==65.3.1
pillow==11.1.0
certifi==2024.12.14
asgiref==3.8.1
attrs==25.1.0
chardet==4.0.0
idna==2.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==1.26.20
channels==4.2.0
channels-redis==4.2.1
daphne==4.0.0
jwt==1.3.1
pillow==11.1.0
pyotp==2.9.0
qrcode==8.0.0
stripe==11.4.1
